Started rework of upkeep system;
* Since the start of Kingdoms, upkeep has always been a flat number that can be added or subtracted. This never gave me a lot of flexibility in managing a kingdom's upkeep, nor did it really give any indication of what was costing a kingdom the most.

I've come up with a new system: Upkeep Objects. Basically every aspect of a kingdom that has upkeep becomes an upkeep object that is attached to that aspect. Create squad --> squad upkeep object, Add chunks --> claim upkeep object etc. This is just the KAPI portion of it, but this system is still pending implementation across the codebase.